echo/server/dembrane/audio_lightrag/requirements.txt
directus_sdk_py
fastapi
lightrag-hku
nest_asyncio==1.6.0
numpy
openai==1.63.2
pandas
pydantic==2.10.6
pydub==0.25.1
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
Requests==2.32.3
uvicorn==0.34.0
